pyproject.toml
[tool.poetry]
name = "calendar-sync"
version = "0.1.0"
description = "A Flask app to sync calendar data from Word documents"
authors = ["Matthew Mazurek"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "3.0.2"
pytest = "8.0.2"
icalendar = "5.0.12"
python-docx = "0.8.11"
requests = "2.31.0"
mammoth = "^1.11.0"
pydantic = "^2.0.0"

[tool.poetry.scripts]
calendar-sync = "cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
